Rename packages
- added *adapter* level in order to prevent package name clash with
package holding the application side client
- renamed packages to better reflect semantics of clients
