+++ clients/adapter/src/main/java/org/eclipse/hono/adapter/client/command/CommandConsumer.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.client.upstream;
+package org.eclipse.hono.adapter.client.command;
 
 import io.opentracing.SpanContext;
 import io.vertx.core.Future;
