+++ clients/adapter/src/main/java/org/eclipse/hono/adapter/client/command/CommandResponse.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.client.upstream;
+package org.eclipse.hono.adapter.client.command;
 
 import java.util.Objects;
 import java.util.function.Predicate;
